footer redesign
