+++ src/app/globals.css
@@ -23,6 +23,7 @@ body {
   align-items: center;
   justify-content: space-between;
   overflow-x: hidden;
+  font-family: var(--font-family);
 }
 
 @layer utilities {
