+++ src/app/product/create/page.tsx
@@ -7,7 +7,7 @@ import {
   ProductCategory,
   ProductCreate,
   ProductType,
-} from "../types";
+} from "../../types";
 import Page from "@/components/Page/Page";
 import Input from "@/components/Input/Input";
 import Button from "@/components/Button/Button";
