+++ src/data/home.ts
@@ -1,0 +1,21 @@
+export const homeHero = {
+  title: "The furniture brand for the future, with timeless designs",
+  button: "View collection",
+  text: "A new era in eco friendly furniture with Avelon, the French luxury retail brand with nice fonts, tasteful colors and a beautiful way to display things digitally using modern web technologies.",
+  imgSrc: "/hero-image.png",
+};
+
+export const homeEmailSignUp = {
+  title: "Join the club and get the benefits",
+  text: "Sign up for our newsletter and receive exclusive offers on new ranges, sales, pop up stores and more",
+};
+
+export const homeInfo = {
+  title: "From a studio in London to a global brand with over 400 outlets",
+  texts: [
+    "When we started Avion, the idea was simple. Make high quality furniture affordable and available for the mass market.",
+    "Handmade, and lovingly crafted furniture and homeware is what we live, breathe and design so our Chelsea boutique become the hotbed for the London interior design community.",
+  ],
+
+  button: "Get in touch",
+};
