+++ src/features/Features/Features.tsx
@@ -1,7 +1,41 @@
 import Container from "@/components/Container/Container";
 import Icon from "@/components/Icon/Icon";
 import Typography from "@/components/Typography/Typography";
-import { homeFeatures } from "@/data/home";
+
+import DeliveryIcon from "@/assets/delivery.svg";
+import CheckmarkIcon from "@/assets/checkmark-outline.svg";
+import PurchaseIcon from "@/assets/purchase.svg";
+import SproutIcon from "@/assets/sprout.svg";
+
+export const homeFeatures = {
+  title: "What makes our brand different",
+  features: [
+    {
+      id: "1",
+      title: "Next day as standard",
+      text: "Order before 3pm and get your order the next day as standard",
+      icon: DeliveryIcon,
+    },
+    {
+      id: "2",
+      title: "Made by true artisans",
+      text: "Handmade crafted goods made with real passion and craftmanship",
+      icon: CheckmarkIcon,
+    },
+    {
+      id: "3",
+      title: "Unbeatable prices",
+      text: "For our materials and quality you won’t find better prices anywhere",
+      icon: PurchaseIcon,
+    },
+    {
+      id: "4",
+      title: "Recycled packaging",
+      text: "We use 100% recycled packaging to ensure our footprint is manageable",
+      icon: SproutIcon,
+    },
+  ],
+};
 
 function Features() {
   return (
