+++ src/features/Sorting/Sorting.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { SortBy, SortOrder } from "@/app/product/types";
+import { SortBy, SortOrder } from "@/app/types";
 import Typography from "@/components/Typography/Typography";
 import { useState } from "react";
 import ArrowIcon from "@/assets/arrow-up.svg";
